fixed (#752)
