+++ apps/mail/hooks/use-email-aliases.ts
@@ -1,13 +1,18 @@
 'use client';
 
-import type { EmailAlias } from '../actions/email-aliases';
 import useSWRImmutable from 'swr/immutable';
 
 const fetcher = (url: string) => fetch(url).then((res) => res.json());
 
+export type EmailAlias = {
+  email: string;
+  name?: string;
+  primary?: boolean;
+};
+
 export function useEmailAliases() {
   const { data, error, isLoading, mutate } = useSWRImmutable<EmailAlias[]>(
-    '/api/driver/email-aliases',
+    '/api/v1/email-aliases',
     fetcher,
   );
